Adding some alerts

+++ src/components/Product.js
@@ -1,5 +1,6 @@
 import React from "react";
 import { Link } from "react-router-dom";
+import Swal from "sweetalert2";
 
 import { useDispatch } from "react-redux";
 import { deleteProductAction } from "../actions/actionsProduct";
@@ -10,7 +11,19 @@ export const Product = ({ product }) => {
 	const dispatch = useDispatch();
 
 	const confirmDeleteProduct = (id) => {
-		dispatch(deleteProductAction(id));
+		Swal.fire({
+			title: "Are you sure?",
+			text: "You won't be able to revert this!",
+			icon: "warning",
+			showCancelButton: true,
+			confirmButtonColor: "#3085d6",
+			cancelButtonColor: "#d33",
+			confirmButtonText: "Yes, delete it!",
+		}).then((result) => {
+			if (result.value) {
+				dispatch(deleteProductAction(id));
+			}
+		});
 	};
 
 	return (
